fix: getuidmap
removed unused const and added tests
