+++ getuidmap/kwok_container_template.tmpl
